refactor: move out device registers definition
