+++ crates/device/Cargo.toml
@@ -7,5 +7,6 @@ edition = "2021"
 
 [dependencies]
 cpu = { version = "0.1.0", path = "../cpu" }
+paste = "1.0.14"
 small-std = { version = "0.1.0", path = "../small-std" }
 tock-registers = "0.9.0"
